Remove dead core gitignore helper
