+++ packages/core/package.json
@@ -41,8 +41,7 @@
     "@ladybugdb/core": "^0.15.3",
     "@tree-sitter-grammars/tree-sitter-kotlin": "1.1.0",
     "@tree-sitter-grammars/tree-sitter-lua": "0.4.1",
-    "ignore": "^7.0.5",
-    "minimatch": "^10.2.5",
+		"minimatch": "^10.2.5",
     "tree-sitter": "^0.25.0",
     "tree-sitter-c": "0.24.1",
     "tree-sitter-c-sharp": "0.23.1",
