+++ packages/extension/tests/acceptance/graphView/workspace.ts
@@ -1,8 +1,8 @@
 import fs from 'node:fs';
 import os from 'node:os';
 import path from 'node:path';
+import ignore, { type Ignore } from 'ignore';
 import { DEFAULT_INCLUDE, DEFAULT_MAX_FILES } from '../../../../core/src/discovery/file/defaults';
-import { loadGitignore } from '../../../../core/src/discovery/gitignore';
 import {
   DEFAULT_EXCLUDE,
   matchesAnyPattern,
@@ -414,6 +414,18 @@ async function discoverAcceptanceWorkspaceFiles(
   });
 
   return files.sort();
+}
+
+function loadGitignore(rootPath: string): Ignore | null {
+  const gitignorePath = path.join(rootPath, '.gitignore');
+
+  if (!fs.existsSync(gitignorePath)) {
+    return null;
+  }
+
+  const gitignore = ignore();
+  gitignore.add(fs.readFileSync(gitignorePath, 'utf-8'));
+  return gitignore;
 }
 
 async function walkAcceptanceDirectory(
